# openapi.json
{"openapi":"3.1.0","info":{"title":"StockIt API","version":"0.1.0"},"paths":{"/api/health":{"get":{"summary":"Healthcheck","operationId":"healthcheck_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Healthcheck Api Health Get"}}}}}}},"/api/auth/register":{"post":{"summary":"Register","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/market/featured":{"get":{"summary":"Featured Stocks","operationId":"featured_stocks_api_market_featured_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response Featured Stocks Api Market Featured Get"}}}}}}},"/api/search":{"get":{"summary":"Search Stock","operationId":"search_stock_api_search_get","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string","default":"","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"string"}},"title":"Response Search Stock Api Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/stock/{symbol}":{"get":{"summary":"Stock Details","operationId":"stock_details_api_stock__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/predict/{symbol}":{"get":{"summary":"Predict Price","operationId":"predict_price_api_predict__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PredictionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/portfolio":{"get":{"summary":"List Portfolio","operationId":"list_portfolio_api_portfolio_get","parameters":[{"name":"x-user-id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PortfolioHoldingRead"},"title":"Response List Portfolio Api Portfolio Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Portfolio Holding","operationId":"create_portfolio_holding_api_portfolio_post","parameters":[{"name":"x-user-id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioHoldingCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PortfolioHoldingRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/watchlist":{"get":{"summary":"List Watchlist","operationId":"list_watchlist_api_watchlist_get","parameters":[{"name":"x-user-id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WatchlistItemRead"},"title":"Response List Watchlist Api Watchlist Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create Watchlist Item","operationId":"create_watchlist_item_api_watchlist_post","parameters":[{"name":"x-user-id","in":"header","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"X-User-Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistItemCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WatchlistItemRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AuthResponse":{"properties":{"user":{"$ref":"#/components/schemas/UserRead"}},"type":"object","required":["user"],"title":"AuthResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IndicatorSeries":{"properties":{"sma_20":{"items":{"additionalProperties":{"anyOf":[{"type":"number"},{"type":"string"}]},"type":"object"},"type":"array","title":"Sma 20"},"rsi_14":{"items":{"additionalProperties":{"anyOf":[{"type":"number"},{"type":"string"}]},"type":"object"},"type":"array","title":"Rsi 14"}},"type":"object","required":["sma_20","rsi_14"],"title":"IndicatorSeries"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":6,"title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"OhlcPoint":{"properties":{"time":{"type":"string","title":"Time"},"open":{"type":"number","title":"Open"},"high":{"type":"number","title":"High"},"low":{"type":"number","title":"Low"},"close":{"type":"number","title":"Close"},"volume":{"type":"number","title":"Volume"}},"type":"object","required":["time","open","high","low","close","volume"],"title":"OhlcPoint"},"PortfolioHoldingCreate":{"properties":{"symbol":{"type":"string","title":"Symbol"},"quantity":{"type":"number","title":"Quantity"},"buy_price":{"type":"number","title":"Buy Price"}},"type":"object","required":["symbol","quantity","buy_price"],"title":"PortfolioHoldingCreate"},"PortfolioHoldingRead":{"properties":{"symbol":{"type":"string","title":"Symbol"},"quantity":{"type":"number","title":"Quantity"},"buy_price":{"type":"number","title":"Buy Price"},"id":{"type":"integer","title":"Id"},"current_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Current Price"},"market_value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Market Value"},"cost_basis":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cost Basis"},"unrealized_gain":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Unrealized Gain"},"unrealized_gain_percent":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Unrealized Gain Percent"},"recommendation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Recommendation"},"recommendation_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Recommendation Reason"}},"type":"object","required":["symbol","quantity","buy_price","id"],"title":"PortfolioHoldingRead"},"PredictionResponse":{"properties":{"symbol":{"type":"string","title":"Symbol"},"model":{"type":"string","title":"Model"},"predicted_price":{"type":"number","title":"Predicted Price"},"latest_close":{"type":"number","title":"Latest Close"},"confidence_note":{"type":"string","title":"Confidence Note"}},"type":"object","required":["symbol","model","predicted_price","latest_close","confidence_note"],"title":"PredictionResponse"},"RegisterRequest":{"properties":{"name":{"type":"string","maxLength":120,"minLength":2,"title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","maxLength":128,"minLength":6,"title":"Password"}},"type":"object","required":["name","email","password"],"title":"RegisterRequest"},"StockQuote":{"properties":{"symbol":{"type":"string","title":"Symbol"},"company_name":{"type":"string","title":"Company Name"},"exchange":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Exchange"},"currency":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Currency"},"current_price":{"type":"number","title":"Current Price"},"previous_close":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Previous Close"},"change":{"type":"number","title":"Change"},"change_percent":{"type":"number","title":"Change Percent"},"market_cap":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Market Cap"},"sector":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"}},"type":"object","required":["symbol","company_name","current_price","change","change_percent"],"title":"StockQuote"},"StockResponse":{"properties":{"quote":{"$ref":"#/components/schemas/StockQuote"},"ohlc":{"items":{"$ref":"#/components/schemas/OhlcPoint"},"type":"array","title":"Ohlc"},"indicators":{"$ref":"#/components/schemas/IndicatorSeries"}},"type":"object","required":["quote","ohlc","indicators"],"title":"StockResponse"},"UserRead":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["id","name","email"],"title":"UserRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WatchlistItemCreate":{"properties":{"symbol":{"type":"string","title":"Symbol"}},"type":"object","required":["symbol"],"title":"WatchlistItemCreate"},"WatchlistItemRead":{"properties":{"symbol":{"type":"string","title":"Symbol"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["symbol","id"],"title":"WatchlistItemRead"}}}}